+++ .idea/compiler.xml
@@ -8,7 +8,7 @@
         <processorPath useClasspath="false">
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.projectlombok/lombok/1.18.32/17d46b3e205515e1e8efd3ee4d57ce8018914163/lombok-1.18.32.jar" />
         </processorPath>
-        <module name="com.apple.shop.main" />
+        <module name="shop.shop.main~1" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
