+++ shop/src/main/java/com/apple/shop/Entity/blog.java
@@ -1,4 +1,4 @@
-package com.apple.shop;
+package com.apple.shop.Entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
