+++ shop/src/main/java/com/apple/shop/ExceptionHandler/MyExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.apple.shop;
+package com.apple.shop.ExceptionHandler;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
