--- shop/src/main/java/com/apple/shop/ItemService.java
@@ -1,32 +1,0 @@
-package com.apple.shop;
-
-import lombok.AllArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.util.List;
-import java.util.Optional;
-
-@Service
-@AllArgsConstructor
-public class ItemService {
-
-    private final ItemRepository itemRepository;
-
-    public void saveItem(String title, Integer price) {
-        Item item = new Item();
-        item.setTitle(title);
-        item.setPrice(price);
-        itemRepository.save(item);
-    }
-
-    public List<Item> findAll() {
-        return itemRepository.findAll();
-    }
-
-    public Optional<Item> findById(Long id) {
-        return itemRepository.findById(id);
-    }
-
-}
