+++ shop/src/main/java/com/apple/shop/Repository/ItemRepository.java
@@ -1,6 +1,7 @@
-package com.apple.shop;
+package com.apple.shop.Repository;
 
 
+import com.apple.shop.Entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemRepository extends JpaRepository<Item,Long> {
