+++ shop/src/main/java/com/apple/shop/Repository/blogRepository.java
@@ -1,6 +1,7 @@
-package com.apple.shop;
+package com.apple.shop.Repository;
 
 
+import com.apple.shop.Entity.blog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface blogRepository extends JpaRepository<blog,Long> {
