+++ shop/src/main/resources/templates/edit.html
@@ -10,14 +10,11 @@
 </head>
 <body>
     <div th:replace="~{ nav.html::navbar }"></div>
-    <div class="card" th:each="i : ${items}">
-        <img src="https://placehold.co/300">
-        <div>
-            <a th:href="@{/detail/{id}(id=${i.id})}"><h4 th:text="${i.title}">바지</h4></a>
-            <p th:text="${i.price + '원'}">7억</p>
-            <a th:href="@{/edit/{id}(id=${i.id})}">✏️</a>
-        </div>
-
-    </div>
+    <form action="/edit" method="POST">
+        <input th:value="${data.id}" name="id" style="display:none">
+        <input th:value="${data.title}" name="title">
+        <input th:value="${data.price}" name="price">
+        <button type="submit">전송</button>
+    </form>
 </body>
 </html>
